requirements.txt
# Core Airflow
apache-airflow==2.7.1
apache-airflow-providers-postgres==5.6.0

# dbt
dbt-postgres==1.6.0
dbt-core==1.6.0

# Database
psycopg2-binary==2.9.7
sqlalchemy==1.4.49

# Data processing
pandas==2.0.3
numpy==1.26.4
requests==2.31.0

# Configuration
python-dotenv==1.0.0

# Development
pytest==7.4.0
black==23.7.0
flake8==6.0.0